Extract methods to the mailchimp client helper
